requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
pylint==3.0.3 